requirements.txt
pyrogram
tgcrypto
google-generativeai
yt-dlp